requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
backpack==0.1
backpack-for-pytorch==1.1.1
beautifulsoup4==4.9.2
bleach==3.2.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dm-haiku @ git+https://github.com/deepmind/dm-haiku@bfa32d5ac3b4dd9a837d74d26ae4e8cb01578e10
dm-tree==0.1.5
docutils==0.16
entrypoints==0.3
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
GPUtil==1.4.0
gpuutils==0.0.2
gradcnn @ git+https://github.com/owkin/grad-cnns.git@b0a9e3bb16f6a2358d3d8e9c936d8d308a648476#subdirectory=code
grpcio==1.32.0
gviz-api==1.9.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jax @ git+https://github.com/google/jax.git@cf1fefcdc65872786e7ed8aafdade35dd4fe8e6c
jaxlib @ https://storage.googleapis.com/jax-releases/cuda110/jaxlib-0.1.55-cp38-none-manylinux2010_x86_64.whl
jedi==0.17.2
jeepney==0.4.3
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.4.0
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
more-itertools==8.5.0
mpmath==1.1.0
multidict==4.7.6
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
nltk==3.5
notebook==6.1.4
numpy==1.18.5
oauthlib==3.1.0
opacus @ git+https://github.com/pytorch/opacus.git@1dc5feb9928ea2962fa751fd996902cdcea1109a
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.0.2
python-dateutil==2.8.1
pytz==2020.1
pyvacy==0.0.32
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
readme-renderer==26.0
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
scipy==1.5.2
seaborn==0.11.0
SecretStorage==3.1.2
Send2Trash==1.5.0
sentencepiece==0.1.91
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
tb-nightly==2.4.0a20200930
tensorboard==2.3.0
tensorboard-plugin-profile==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
-e git+https://github.com/tensorflow/privacy@20d0b884bab21010318dd677cd0bf0768c371849#egg=tensorflow_privacy
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-estimator-nightly==2.4.0.dev2020093001
tf-nightly-gpu==2.4.0.dev20201005
toml==0.10.1
torch==1.6.0
torchcsprng==0.1.2
torchtext==0.7.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.49.0
traitlets==5.0.4
twine==3.2.0
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yapf==0.30.0
yarl==1.6.0
zipp==3.1.0
